+++ database/migrations/2021_01_16_133919_create_staff_awards_table.php
@@ -14,7 +14,7 @@ class CreateStaffAwardsTable extends Migration
     public function up()
     {
         Schema::create('staff_awards', function (Blueprint $table) {
-            $table->unsignedInteger('award_id')->default(NULL)->comment('ID');
+            $table->unsignedBigInteger('award_id')->default(NULL)->comment('ID');
             $table->string('award_name')->default(NULL)->comment('Наименование');
             $table->string('award_type')->default(NULL)->comment('Тип (1.1.6.3.1)');
             $table->string('award_type_status')->default(NULL)->comment('Тип награды (1.1.6.3.2)');
