+++ database/migrations/2021_01_16_134051_create_staff_deleteds_table.php
@@ -14,7 +14,7 @@ class CreateStaffDeletedsTable extends Migration
     public function up()
     {
         Schema::create('staff_deleteds', function (Blueprint $table) {
-            $table->unsignedInteger('deleted_id')->default(NULL)->comment('');
+            $table->unsignedBigInteger('deleted_id')->default(NULL)->comment('');
             $table->boolean('deleted_status')->default(FALSE)->comment('Удалить без приказа(Check) (удаляет дубликат или ошибочные данные)');
             $table->binary('delete_date')->default(NULL)->comment('Дата окончания отпуска');
             $table->string('delete_dismissal_order_number')->default(NULL)->comment('Номер приказа об увольнении');
